renamed config module to configuration
naming consistency only
