+++ src/endpoint/simple-fastapi-container/src/api/configuration.py
